nsimplex {combinat}	R Documentation

Computes the number of points on a (p, n)-simplex lattice

Description

Computes the number of points on a (p, n)-simplex lattice; that is, the number of p-part compositions of n. This gives the number of points in the support space of a Multinomial(n, q) distribution, where p == length(q).

Arguments p and n are replicated as necessary to have the length of the longer of them.

Usage

nsimplex(p, n)

Arguments

p
vector of integers

n
vector of integers

Value

integer

Examples

nsimplex(3,5)

[Package combinat version 0.0-8 Index]