xsimplex {combinat}	R Documentation

Generates all points on a (p,n) simplex lattice (i.e. a p-part composition of n).

Description

Generates all points on a p,n simplex lattice (i.e. a p-part composition of n). Each point is represented as x, a p-dimensional vector of nonnegative integers that sum to n. If argument "fun" is not null, applies a function given by the argument to each point. If simplify is FALSE, returns a list; else returns a vector or an array. "..." are passed unchanged to function given by argument fun, if any.

Usage

xsimplex(p, n, fun=NULL, simplify=TRUE, ...)

Arguments

p
first parameter of lattice description

n
second parameter of lattice description

fun
function to be applied pointwise

simplify
logical: if FALSE, value is a list, otherwise a vector or array

...
parameters to be passed to fun

Examples

#Compute Multinomial(n = 4, pi = rep(1/3, 3)) p.f.:
   xsimplex(3, 4, dmnom, prob=1/3)

[Package combinat version 0.0-8 Index]